refact: 타입 위치 수정 && fetch 부분들 전부 hooks로 이동
