--- src/apis/color/index.ts
@@ -1,3 +1,0 @@
-export * from "./getIntColorInfos";
-export * from "./getExtColorInfos";
-export * from "./getChangeableCarModelsWithTrim";
