--- src/apis/model/index.ts
@@ -1,5 +1,0 @@
-export * from "./getCarInfo";
-export * from "./getCarInfos";
-export * from "./getModelInfo";
-export * from "./getModelFilters";
-export * from "./getTrimInfos";
