+++ src/components/MainPage/CarList.tsx
@@ -3,6 +3,7 @@ import { carListSelector } from "@/stores/carState";
 import { useRecoilValue } from "recoil";
 import { useNavigate } from "react-router-dom";
 import { ROUTE_PATH } from "@/Router";
+import { useImageUrl } from "@/hooks/useImageUrl";
 
 export const CarList = () => {
   const carList = useRecoilValue(carListSelector);
@@ -21,7 +22,7 @@ export const CarList = () => {
               key={car.carCode}
               onClick={() => handleCarInfoClick(car.carCode)}
             >
-              <img src={import.meta.env.VITE_BACKEND_URL + car.carImagePath} />
+              <img src={useImageUrl(car.carImagePath)} />
               <strong>{car.carName}</strong>
               <p>{(car.carLowPrice / 10000).toLocaleString()}만원 ~</p>
             </CarInfoDiv>
