+++ src/hooks/useImageUrl.ts
@@ -1,0 +1,3 @@
+export const useImageUrl = (imagePath: string): string => {
+  return import.meta.env.VITE_BACKEND_URL + imagePath;
+};
