+++ src/services/option.ts
@@ -1,0 +1,6 @@
+import { OptionInfo } from "@/types/option";
+import { apiInstance } from "./instance";
+
+export const getOptions = async (modelCode: string): Promise<OptionInfo[]> => {
+  return await apiInstance.get<OptionInfo[]>(`/option/?modelCode=${modelCode}`);
+};
