+++ src/types/option.ts
@@ -1,0 +1,13 @@
+export type OptionInfo = {
+  optionId: number;
+  optionCode: string;
+  optionName: string;
+  optionPrice: number;
+  optionImagePath: string;
+  optionTypeName: string;
+  isSelectable: boolean;
+};
+
+export type ExtendedOptionInfo = OptionInfo & {
+  isSelected: boolean;
+};
